+++ record_macos/macos/Classes/RecordMacosPlugin.swift
@@ -2,16 +2,15 @@ import Cocoa
 import FlutterMacOS
 import AVFoundation
 
-public class RecordMacosPlugin: NSObject, FlutterPlugin, AVAudioRecorderDelegate {
+public class RecordMacosPlugin: NSObject, FlutterPlugin, AVCaptureFileOutputRecordingDelegate {
   public static func register(with registrar: FlutterPluginRegistrar) {
     let channel = FlutterMethodChannel(name: "com.llfbandit.record", binaryMessenger: registrar.messenger)
     let instance = RecordMacosPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
   }
 
-  var isRecording = false
-  var isPaused = false
-  var audioRecorder: AVAudioRecorder?
+  var audioSession: AVCaptureSession?
+  var audioOutput: AVCaptureAudioFileOutput?
   var path: String?
   var maxAmplitude:Float = -160.0;
 
@@ -33,6 +32,8 @@ public class RecordMacosPlugin: NSObject, FlutterPlugin, AVAudioRecorderDelegate
           encoder: args["encoder"] as! String,
           bitRate: args["bitRate"] as? Int ?? 128000,
           samplingRate: args["samplingRate"] as? Int ?? 44100,
+          numChannels: args["numChannels"] as? Int ?? 2,
+          device: args["device"] as? [String : Any],
           result: result)
         break
       case "stop":
@@ -40,12 +41,15 @@ public class RecordMacosPlugin: NSObject, FlutterPlugin, AVAudioRecorderDelegate
         break
       case "pause":
         pause(result)
+        break
       case "resume":
         resume(result)
+        break
       case "isPaused":
-        result(isPaused)
+        result(audioOutput?.isRecordingPaused ?? false)
+        break
       case "isRecording":
-        result(isRecording)
+        result(audioOutput?.isRecording ?? false)
         break
       case "hasPermission":
         hasPermission(result)
@@ -58,6 +62,15 @@ public class RecordMacosPlugin: NSObject, FlutterPlugin, AVAudioRecorderDelegate
         let encoder = args["encoder"] as! String
         let settings = getEncoderSettings(encoder)
         result(settings != nil)
+        break
+      case "listInputDevices":
+      let devs = listInputDevices().map { (device) -> [String : Any] in
+        return [
+          "id": device.uniqueID,
+          "label": device.localizedName,
+        ]
+      }
+        result(devs)
         break
       case "dispose":
         dispose(result)
@@ -73,9 +86,6 @@ public class RecordMacosPlugin: NSObject, FlutterPlugin, AVAudioRecorderDelegate
       case .authorized:
         result(true)
         break
-      case .denied:
-        result(false)
-        break
       case .notDetermined:
         AVCaptureDevice.requestAccess(for: .audio) { allowed in
           DispatchQueue.main.async {
@@ -84,28 +94,61 @@ public class RecordMacosPlugin: NSObject, FlutterPlugin, AVAudioRecorderDelegate
         }
         break
       default:
+        result(false)
         break
     }
   }
 
-  fileprivate func start(path: String, encoder: String, bitRate: Int, samplingRate: Int, result: @escaping FlutterResult) {
+  fileprivate func start(path: String, encoder: String, bitRate: Int, samplingRate: Int, numChannels: Int, device: [String: Any]?, result: @escaping FlutterResult) {
     stopRecording()
 
-    let settings = getSettings(encoder: encoder, bitRate: bitRate, samplingRate: samplingRate)
-
-    do {
-      let url = URL(string: path) ?? URL(fileURLWithPath: path)
-      audioRecorder = try AVAudioRecorder(url: url, settings: settings)
-      audioRecorder!.delegate = self
-      audioRecorder!.isMeteringEnabled = true
-      audioRecorder!.record()
-
-      isRecording = true
-      isPaused = false
-      result(nil)
-    } catch {
-      result(FlutterError(code: "", message: "Failed to start recording", details: error))
+    audioSession = AVCaptureSession()
+    guard let audioSession = audioSession else {
+      result(FlutterError(code: "-1", message: "Failed to start recording", details: "Audio session is nil."))
+      return
     }
+
+    let dev: AVCaptureInput?
+    do {
+      dev = try getInputDevice(device: device)
+    } catch {
+      result(FlutterError(code: "-2", message: "Failed to start recording", details: "\(error)"))
+      return
+    }
+    
+    guard let dev = dev else {
+      result(FlutterError(code: "-3", message: "Failed to start recording", details: "Input device not found from available list."))
+      return
+    }
+    guard audioSession.canAddInput(dev) else {
+      result(FlutterError(code: "-4", message: "Failed to start recording", details: "Input device cannot be added to the capture session."))
+      return
+    }
+
+    audioSession.beginConfiguration()
+
+    // Add input device
+    audioSession.addInput(dev)
+    // Add output
+    audioOutput = AVCaptureAudioFileOutput()
+    audioOutput!.audioSettings = getSettings(
+      encoder: encoder,
+      bitRate: bitRate,
+      samplingRate: samplingRate,
+      numChannels: numChannels)
+    audioSession.addOutput(audioOutput!)
+
+    audioSession.commitConfiguration()
+
+    audioSession.startRunning()
+    
+    audioOutput!.startRecording(
+      to: URL(string: path) ?? URL(fileURLWithPath: path),
+      outputFileType: AVFileType.m4a,
+      recordingDelegate: self
+    )
+
+    result(nil)
   }
 
   fileprivate func stop(_ result: @escaping FlutterResult) {
@@ -114,15 +157,13 @@ public class RecordMacosPlugin: NSObject, FlutterPlugin, AVAudioRecorderDelegate
   }
     
   fileprivate func pause(_ result: @escaping FlutterResult) {
-    audioRecorder?.pause()
-    isPaused = true
+    audioOutput?.pauseRecording()
     result(nil)
   }
     
   fileprivate func resume(_ result: @escaping FlutterResult) {
-    if isPaused {
-      audioRecorder?.record()
-      isPaused = false
+    if audioOutput?.isRecordingPaused ?? false {
+      audioOutput?.resumeRecording()
     }
     
     result(nil)
@@ -136,30 +177,38 @@ public class RecordMacosPlugin: NSObject, FlutterPlugin, AVAudioRecorderDelegate
   fileprivate func getAmplitude(_ result: @escaping FlutterResult) {
     var amp = ["current" : -160.0, "max" : -160.0] as [String : Float]
 
-    if isRecording {
-      audioRecorder?.updateMeters()
+    if let audioOutput = audioOutput {
+//      var channels = 0
+//      var decibels: Float = 0.0
+//      for connection in audioOutput.connections {
+//        for audioChannel in connection.audioChannels {
+//          decibels += audioChannel.averagePowerLevel
+//          channels += 1
+//        }
+//      }
+//
+//      decibels /= Float(channels)
+//      let current = pow(10.0, 0.05 * decibels) * 20.0
       
-      guard let current = audioRecorder?.averagePower(forChannel: 0) else {
-        result(amp)
-        return
-      }
+      let current = audioOutput.connections.first?.audioChannels.first?.averagePowerLevel
+      if let current = current {
+        if (current > maxAmplitude) {
+          maxAmplitude = current
+        }
 
-      if (current > maxAmplitude) {
-        maxAmplitude = current
+        amp["current"] = current
+        amp["max"] = maxAmplitude
       }
-
-      amp["current"] = current
-      amp["max"] = maxAmplitude
     }
 
     result(amp)
   }
 
   fileprivate func stopRecording() {
-    audioRecorder?.stop()
-    audioRecorder = nil
-    isRecording = false
-    isPaused = false
+    audioOutput?.stopRecording()
+    audioOutput = nil
+    audioSession?.stopRunning()
+    audioSession = nil
     maxAmplitude = -160.0
   }
 
@@ -167,12 +216,21 @@ public class RecordMacosPlugin: NSObject, FlutterPlugin, AVAudioRecorderDelegate
     stopRecording()
     result(path)
   }
+  
+  public func fileOutput(_ output: AVCaptureFileOutput,
+                         didFinishRecordingTo outputFileURL: URL,
+                         from connections: [AVCaptureConnection],
+                         error: Error?) {
+    if let error = error {
+      print(error)
+    }
+  }
 
-  fileprivate func getSettings(encoder: String, bitRate: Int, samplingRate: Int) -> [String : Any] {
+  fileprivate func getSettings(encoder: String, bitRate: Int, samplingRate: Int, numChannels: Int) -> [String : Any] {
     let settings = [
       AVEncoderBitRateKey: bitRate,
       AVSampleRateKey: samplingRate,
-      AVNumberOfChannelsKey: 2,
+      AVNumberOfChannelsKey: numChannels,
       AVEncoderAudioQualityKey: AVAudioQuality.high.rawValue
     ] as [String : Any]
 
@@ -186,7 +244,7 @@ public class RecordMacosPlugin: NSObject, FlutterPlugin, AVAudioRecorderDelegate
   }
 
   // https://developer.apple.com/documentation/coreaudiotypes/coreaudiotype_constants/1572096-audio_data_format_identifiers
-  fileprivate func getEncoderSettings(_ encoder: String) -> [String : Any]? {    
+  fileprivate func getEncoderSettings(_ encoder: String) -> [String : Any]? {
     switch(encoder) {
     case "aacEld":
       return [AVFormatIDKey : Int(kAudioFormatMPEG4AAC_ELD)]
@@ -215,5 +273,37 @@ public class RecordMacosPlugin: NSObject, FlutterPlugin, AVAudioRecorderDelegate
     default:
       return nil
     }
+  }
+
+  fileprivate func listInputDevices() -> [AVCaptureDevice] {
+    let discoverySession = AVCaptureDevice.DiscoverySession(
+      deviceTypes: [.builtInMicrophone],
+      mediaType: .audio, position: .unspecified
+    )
+    
+    return discoverySession.devices
+  }
+
+  fileprivate func getInputDevice(device: [String: Any]?) throws -> AVCaptureDeviceInput? {
+    guard let device = device else {
+      // try to select default device
+      let defaultDevice = AVCaptureDevice.default(for: .audio)
+      guard let defaultDevice = defaultDevice else {
+        return nil
+      }
+
+      return try AVCaptureDeviceInput(device: defaultDevice)
+    }
+
+    // find the given device
+    let devs = listInputDevices()
+    let captureDev = devs.first { dev in
+      dev.uniqueID == device["id"] as! String
+    }
+    guard let captureDev = captureDev else {
+      return nil
+    }
+    
+    return try AVCaptureDeviceInput(device: captureDev)
   }
 }