+++ record_linux/CHANGELOG.md
@@ -1,2 +1,7 @@
+## 0.2.0
+- feat: Add input devices listing.
+- feat: Add number of channels.
+- core: Update fmedia to version 1.28.
+
 ## 0.1.0
 * core: Initial release